feat (front): criando a logica de apresentacao do modal
