+++ front/src/renderer/src/views/layouts/PageLayout/components/Aside/Payments/modals/NewPaymentModal/useNewPaymentModal.ts
@@ -1,50 +1,12 @@
-import { zodResolver } from '@hookform/resolvers/zod';
-import { PaymentType } from '@renderer/app/entities/Payment';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { PaymentFormSchema } from '../../PaymentForm/usePaymentFormModal';
 
-export const paymentFormSchema = z.object({
-  type: z.enum(['CREDIT_CARD', 'DEBIT_CARD', 'CASH']),
-  date: z.date(),
-  value: z.number({ required_error: 'O valor é obrigatório.' }),
-});
-
-export type PaymentFormSchema = z.infer<typeof paymentFormSchema>;
-
-export default function useNewPaymentModal() {
-  const {
-      register,
-      handleSubmit: hookFormHandleSubmit,
-      formState: { errors },
-      control,
-      setFocus,
-      watch,
-      setValue,
-    } = useForm<PaymentFormSchema>({
-      resolver: zodResolver(paymentFormSchema),
-      defaultValues: {
-        type: 'DEBIT_CARD',
-        date: new Date(),
-      },
-    });
-
-  const selectedType = watch('type');
-
-  function handleSelectedType(type: PaymentType) {
-    setValue('type', type, { shouldValidate: true });
+export default function useNewPaymentModal(onSuccess: () => void) {
+  async function handleSubmit(data: PaymentFormSchema) {
+    console.log(data.date.toISOString());
+    onSuccess();
   }
 
-  const handleSubmit = hookFormHandleSubmit((data) => {
-    console.log({ data });
-  });
-
   return {
-    errors,
-    control,
-    selectedType,
-    register,
-    setFocus,
-    handleSelectedType,
     handleSubmit,
   };
 }
